Prepare editor for having tools and agents as a sidebar In this PR I'll put all the layout with sidebar under a feature flag for now
