+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/_components/DocumentEditor/Editor/index.tsx
@@ -1,5 +1,17 @@
 'use client'
 
-import { DocumentEditor } from './DocumentEditor'
+import { DocumentEditorContentArea } from './ContentArea'
 
-export default DocumentEditor
+export default function DocumentEditor(props: {
+  freeRunsCount?: number
+  refinementEnabled: boolean
+  showPreview?: boolean
+}) {
+  return (
+    <DocumentEditorContentArea
+      freeRunsCount={props.freeRunsCount}
+      refinementEnabled={props.refinementEnabled}
+      showPreview={props.showPreview}
+    />
+  )
+}
